feature: Interface de réservation - Wilson Rocha
- Ajout des composants de réservation
- Correction du sidebar avec getUserName()
- Mise à jour des services et modèles
- Ajout des modals pour login, register, create book

+++ bibliotheque-frontend/src/app/books-list/books-list.component.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { Books } from '../_model/books'
+import { Books } from '../_model/books';
 import { BooksService } from '../_service/books.service';
+import { ModalService } from '../_service/modal.service';
 
 @Component({
   selector: 'app-books-list',
@@ -12,31 +13,43 @@ export class BooksListComponent implements OnInit {
 
   books: Books[];
 
-  constructor(private booksService: BooksService,
-    private router: Router) { }
+  constructor(
+    private booksService: BooksService,
+    private router: Router,
+    private modalService: ModalService
+  ) {}
 
   ngOnInit(): void {
     this.getBooks();
   }
 
   private getBooks() {
-    this.booksService.getBooksList().subscribe(data =>{
+    this.booksService.getBooksList().subscribe(data => {
       this.books = data;
     });
   }
 
   updateBook(bookId: number) {
-    this.router.navigate(['update-book', bookId ]);
+    this.router.navigate(['update-book', bookId]);
   }
 
-  deleteBook(bookId: number) {
-    this.booksService.deleteBook(bookId).subscribe( data=> {
-      this.getBooks();
+  async deleteBook(bookId: number) {
+    const confirmed = await this.modalService.confirm({
+      title: 'Supprimer le livre',
+      message: 'Êtes-vous sûr de vouloir supprimer ce livre ? Cette action est irréversible.',
+      confirmText: 'Supprimer',
+      cancelText: 'Annuler',
+      type: 'danger'
     });
+
+    if (confirmed) {
+      this.booksService.deleteBook(bookId).subscribe(data => {
+        this.getBooks();
+      });
+    }
   }
 
   bookDetails(bookId: number) {
-    this.router.navigate(['book-details', bookId ]);
+    this.router.navigate(['book-details', bookId]);
   }
-
 }
